+++ tests/tmain.c
@@ -104,6 +104,7 @@ static const tfam_t fam_order[] = {
 
     { "tri", "ttriton.c",   "Triton frontend",          2 },
     { "mlr", "tmlir.c",     "MLIR reader",              2 },
+    { "bir", "tbir.c",      "BIR text frontend",        2 },
 
     { "sfp", "tsoft_fp.c",  "soft float",               2 },
     { "spr", "tsysprint.c", "SYSPRINT",                 2 },